llama-mysql-demo/requirements.txt
mysql-connector-python
llama-index
pymysql
sqlalchemy
python-dotenv